LayoutTests/editing/style/inverse-color-filter-expected.txt

| <div>
|   contenteditable=""
|   data-color="rgb(0, 0, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 255, 255)"
|   <span>
|     style="background-color: rgb(0, 0, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(235, 235, 235)"
|   <span>
|     style="background-color: rgb(25, 25, 25);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(214, 214, 214)"
|   <span>
|     style="background-color: rgb(51, 51, 51);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(192, 192, 192)"
|   <span>
|     style="background-color: rgb(79, 79, 79);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(170, 170, 170)"
|   <span>
|     style="background-color: rgb(106, 106, 106);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(146, 146, 146)"
|   <span>
|     style="background-color: rgb(136, 136, 136);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(122, 122, 122)"
|   <span>
|     style="background-color: rgb(166, 166, 166);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(96, 96, 96)"
|   <span>
|     style="background-color: rgb(199, 199, 199);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(68, 68, 68)"
|   <span>
|     style="background-color: rgb(234, 234, 234);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(35, 35, 35)"
|   <span>
|     style="background-color: rgb(255, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 0, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 54, 74)"
|   <span>
|     style="background-color: rgb(232, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 30, 87)"
|   <span>
|     style="background-color: rgb(220, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(17, 5, 59)"
|   <span>
|     style="background-color: rgb(255, 252, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(46, 7, 62)"
|   <span>
|     style="background-color: rgb(255, 247, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(60, 8, 27)"
|   <span>
|     style="background-color: rgb(255, 244, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(92, 7, 0)"
|   <span>
|     style="background-color: rgb(255, 207, 198);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(90, 28, 0)"
|   <span>
|     style="background-color: rgb(255, 234, 199);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(88, 52, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 199);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(85, 61, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 202);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(102, 97, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 177);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(79, 85, 3)"
|   <span>
|     style="background-color: rgb(255, 255, 201);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(38, 62, 15)"
|   <span>
|     style="background-color: rgb(255, 255, 236);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 77, 101)"
|   <span>
|     style="background-color: rgb(200, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 46, 122)"
|   <span>
|     style="background-color: rgb(180, 237, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(26, 10, 83)"
|   <span>
|     style="background-color: rgb(245, 225, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(69, 14, 89)"
|   <span>
|     style="background-color: rgb(255, 216, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(86, 16, 41)"
|   <span>
|     style="background-color: rgb(255, 211, 243);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(131, 17, 0)"
|   <span>
|     style="background-color: rgb(255, 168, 147);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(124, 42, 0)"
|   <span>
|     style="background-color: rgb(255, 206, 153);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(122, 74, 0)"
|   <span>
|     style="background-color: rgb(255, 245, 153);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(120, 87, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 154);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(141, 134, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 122);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(111, 118, 8)"
|   <span>
|     style="background-color: rgb(255, 255, 156);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(56, 87, 26)"
|   <span>
|     style="background-color: rgb(240, 255, 203);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 109, 143)"
|   <span>
|     style="background-color: rgb(150, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 66, 170)"
|   <span>
|     style="background-color: rgb(125, 207, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(44, 19, 118)"
|   <span>
|     style="background-color: rgb(216, 184, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(97, 23, 124)"
|   <span>
|     style="background-color: rgb(255, 175, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(121, 26, 62)"
|   <span>
|     style="background-color: rgb(255, 168, 213);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(181, 26, 0)"
|   <span>
|     style="background-color: rgb(255, 113, 81);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(173, 62, 0)"
|   <span>
|     style="background-color: rgb(255, 165, 88);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(169, 104, 0)"
|   <span>
|     style="background-color: rgb(255, 219, 89);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(167, 123, 0)"
|   <span>
|     style="background-color: rgb(255, 243, 90);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(196, 188, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 46);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(154, 166, 14)"
|   <span>
|     style="background-color: rgb(255, 255, 90);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(79, 122, 40)"
|   <span>
|     style="background-color: rgb(205, 255, 156);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 140, 180)"
|   <span>
|     style="background-color: rgb(106, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 86, 214)"
|   <span>
|     style="background-color: rgb(74, 182, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(55, 26, 148)"
|   <span>
|     style="background-color: rgb(186, 150, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(123, 33, 159)"
|   <span>
|     style="background-color: rgb(246, 134, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(154, 36, 79)"
|   <span>
|     style="background-color: rgb(255, 127, 180);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(227, 36, 0)"
|   <span>
|     style="background-color: rgb(255, 65, 20);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(217, 80, 0)"
|   <span>
|     style="background-color: rgb(255, 129, 29);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(213, 132, 0)"
|   <span>
|     style="background-color: rgb(255, 194, 29);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(210, 157, 0)"
|   <span>
|     style="background-color: rgb(255, 227, 31);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(245, 236, 0)"
|   <span>
|     style="background-color: rgb(255, 255, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(195, 209, 23)"
|   <span>
|     style="background-color: rgb(246, 255, 31);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(102, 156, 53)"
|   <span>
|     style="background-color: rgb(173, 240, 111);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 163, 215)"
|   <span>
|     style="background-color: rgb(66, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 97, 255)"
|   <span>
|     style="background-color: rgb(28, 149, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(77, 34, 179)"
|   <span>
|     style="background-color: rgb(168, 114, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(153, 41, 189)"
|   <span>
|     style="background-color: rgb(239, 99, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(185, 45, 93)"
|   <span>
|     style="background-color: rgb(255, 88, 148);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 64, 19)"
|   <span>
|     style="background-color: rgb(255, 40, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 106, 0)"
|   <span>
|     style="background-color: rgb(255, 109, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 170, 0)"
|   <span>
|     style="background-color: rgb(255, 183, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(254, 199, 0)"
|   <span>
|     style="background-color: rgb(255, 218, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 252, 65)"
|   <span>
|     style="background-color: rgb(210, 206, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(217, 235, 55)"
|   <span>
|     style="background-color: rgb(202, 225, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(119, 187, 65)"
|   <span>
|     style="background-color: rgb(138, 223, 71);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(0, 199, 252)"
|   <span>
|     style="background-color: rgb(21, 255, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(58, 136, 254)"
|   <span>
|     style="background-color: rgb(22, 120, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(94, 48, 235)"
|   <span>
|     style="background-color: rgb(107, 50, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(190, 56, 243)"
|   <span>
|     style="background-color: rgb(203, 36, 255);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(230, 59, 122)"
|   <span>
|     style="background-color: rgb(246, 32, 111);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 98, 81)"
|   <span>
|     style="background-color: rgb(207, 11, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 134, 71)"
|   <span>
|     style="background-color: rgb(215, 63, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 180, 63)"
|   <span>
|     style="background-color: rgb(219, 125, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(254, 203, 62)"
|   <span>
|     style="background-color: rgb(218, 154, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 247, 107)"
|   <span>
|     style="background-color: rgb(164, 154, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(228, 239, 101)"
|   <span>
|     style="background-color: rgb(159, 173, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(150, 211, 95)"
|   <span>
|     style="background-color: rgb(110, 186, 41);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(83, 213, 253)"
|   <span>
|     style="background-color: rgb(15, 177, 227);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(116, 167, 254)"
|   <span>
|     style="background-color: rgb(16, 80, 189);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(135, 78, 254)"
|   <span>
|     style="background-color: rgb(95, 24, 244);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(211, 87, 254)"
|   <span>
|     style="background-color: rgb(175, 20, 228);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(237, 113, 158)"
|   <span>
|     style="background-color: rgb(178, 23, 79);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 140, 130)"
|   <span>
|     style="background-color: rgb(149, 5, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 165, 125)"
|   <span>
|     style="background-color: rgb(152, 40, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 198, 119)"
|   <span>
|     style="background-color: rgb(156, 84, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 216, 119)"
|   <span>
|     style="background-color: rgb(154, 105, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 249, 149)"
|   <span>
|     style="background-color: rgb(118, 110, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(235, 243, 143)"
|   <span>
|     style="background-color: rgb(116, 126, 1);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(177, 221, 140)"
|   <span>
|     style="background-color: rgb(79, 134, 33);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(148, 227, 254)"
|   <span>
|     style="background-color: rgb(9, 108, 142);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(168, 198, 254)"
|   <span>
|     style="background-color: rgb(11, 48, 118);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(177, 140, 254)"
|   <span>
|     style="background-color: rgb(62, 16, 159);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(227, 146, 254)"
|   <span>
|     style="background-color: rgb(114, 13, 148);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(244, 164, 192)"
|   <span>
|     style="background-color: rgb(114, 14, 49);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 181, 175)"
|   <span>
|     style="background-color: rgb(95, 3, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 196, 171)"
|   <span>
|     style="background-color: rgb(98, 25, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 217, 168)"
|   <span>
|     style="background-color: rgb(100, 52, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 228, 168)"
|   <span>
|     style="background-color: rgb(99, 65, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 251, 185)"
|   <span>
|     style="background-color: rgb(78, 73, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(242, 247, 183)"
|   <span>
|     style="background-color: rgb(75, 81, 1);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(204, 232, 181)"
|   <span>
|     style="background-color: rgb(51, 86, 23);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(202, 240, 254)"
|   <span>
|     style="background-color: rgb(5, 53, 70);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(212, 227, 254)"
|   <span>
|     style="background-color: rgb(6, 25, 58);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(217, 202, 254)"
|   <span>
|     style="background-color: rgb(27, 8, 73);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(241, 201, 254)"
|   <span>
|     style="background-color: rgb(57, 7, 73);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(249, 211, 224)"
|   <span>
|     style="background-color: rgb(55, 8, 24);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 218, 216)"
|   <span>
|     style="background-color: rgb(47, 0, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 226, 214)"
|   <span>
|     style="background-color: rgb(48, 12, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 236, 213)"
|   <span>
|     style="background-color: rgb(48, 24, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(255, 242, 213)"
|   <span>
|     style="background-color: rgb(48, 31, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(254, 252, 221)"
|   <span>
|     style="background-color: rgb(38, 35, 0);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(248, 250, 219)"
|   <span>
|     style="background-color: rgb(38, 41, 2);"
|     "aaa"
| <div>
|   contenteditable=""
|   data-color="rgb(224, 237, 212)"
|   <span>
|     style="background-color: rgb(34, 51, 19);"
|     "<#selection-anchor>aaa<#selection-focus>"
